Backend/FastAPI/requirements.txt
fastapi[standard]
python-jose
passlib
bcrypt
pymongo
uvicorn